Move GasPriceSelector to ui
