+++ js/src/ui/GasPriceSelector/CustomBar/customBar.js
