+++ js/src/ui/GasPriceSelector/CustomBar/index.js
